services/datagen/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = ["intey <ziexe0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
rabbitmq = "^0.2.0"
fastapi = "^0.60.1"
pydantic = "^1.6.1"
pika = "^1.1.0"
mimesis = "^4.0.0"
uvicorn = "^0.11.6"
kafka-python = "^2.0.1"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
